Make VLSM a record, combine non-VLSMType fields into VLSMMachine (#14)
This changes the definition of VLSM from a nested sigT into
a record, and collapses the two records VLSMSign and VLSMClass
into a single VLSMMachine record because that division was never
useful.
(Separating VLSMType from the rest lets us write definitions like
VLSM_incl that can only be applied to two VLSMs with the same
VLSMType, and it was never helpful to talk about two VLSMs with
the same initial messages and states but different transitions.
The other possible use would be if definitions like
pre_loaded_with_all_messages_vlsm could reuse some of the records,
but VLSMClass dependend on VLSMSign and we never wanted to
change just the transition or valid functions while leaving the
initial states or messages alone).

Unfortunately this somehow changed things so that the extremely
limited form of unification used in the `apply` tactic can no longer
directly apply some lemmas to some goals. For example, lemmas with
conclusions about composite_transition no longer work with the
plain `apply` if the goal was (vtransition (free_composite_vlsm IM)),
even though that expression can reduce to composite_transition.

This can usually be fixed by using `rapply`, which is like
`apply` but based on `refine` and uses a less limited unification
algorithm. I now export it from our Preamble (it's from
Coq.Program.Tactics).
If that fails you can also explicitly supply more arguments.

There is not an `rapply lem in H` form, but the special case of
`apply lem in H;assumption.` can become `revert H;rapply lem`.